Move library config objects out of app module
